Create simple Deskpro client
We need to be able to create tickets in Deskpro, there was no ruby
client for communicating with the Deskpro API available.

This uses 'rest-client' to create an ActiveResource-style interface for
working with the API and helping with client-side validation.

We use the v1 API rather than the newer v2 API since the v2 API has very
few implementations (the offical PHP SDK is based on v1) and because
creating a ticket with a message requires at least 2 requests with v2
API which left open the posibility of an inconsistant state due to no
transaction support.

As a rough example, to implement an Agent 'resource'  you might do:

```
class Agent < Deskpro::Resource
  path: '/agents'
  field :id, Integer
  field :name, String, :match => /[a-zA-Z]/
end
```

You could then interact with the resource using the Deskpro::Client

```
client = Deskpro::Client.new({
  api_key: 'xxx',
  endpoint: 'https://account.deskpro.com'
})
agent = Agent.new(name: 'jeff')
raise "invalid" if !agent.valid?
client.post agent
```

So far only a `Deskpro::Ticket` resource and the `Deskpro::Client#post`
is implemented.