Remove 'register' and 'signup' labels
The deskpro agent assigned to this app in production does not have the
ability to create "labels".

The `register` and `signup` labels do not exist yet were being ignored.

This removes the mention of the ignored labels. If someone wants to add
additional labels they will either have to manually add labels via
Deskpro UI first or give the user permission to create labels.